fix(collision): fixing the collisions
- Stop the jump when a collision is detected above the player
- Enabling platforms movement
- Moving collisions calculations in Game::update() after applying the movements
- Changing collision correction : using 'x' and 'y' values instead of 'moveX' and 'moveY'
- Adding the case in correctSATCollision() where the collision is on both axis
- Renaming camera methods : removing 'camera' in names

+++ src/Game/Platforms/MovingPlatform1D.cpp
@@ -72,8 +72,7 @@ void MovingPlatform1D::setIsMoving(bool state) {
 
 void MovingPlatform1D::applyXaxisMovement(double deltaTime) {
     move = 150; // Add basic movement
-
-    move *= deltaTime; // Apply movement per second
+    move *= static_cast<float>(deltaTime); // Apply movement per second
     move *= speed; // Apply speed
     move *= direction; // Apply direction
 
@@ -94,7 +93,7 @@ void MovingPlatform1D::applyXaxisMovement(double deltaTime) {
 void MovingPlatform1D::applyYaxisMovement(double deltaTime) {
     move = 150; // Add basic movement
 
-    move *= deltaTime; // Apply movement per second
+    move *= static_cast<float>(deltaTime); // Apply movement per second
     move *= speed; // Apply speed
     move *= direction; // Apply direction
